+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Features
 - Added support for grouping shapes together in a Group - @Pr0phet GH-68
 - Added support for calculating the offset and extent on a Slide. - @Pr0phet GH-68
+- Added support for Horizontal bar chart - @rdoepke @Progi1984 GH-58
 
 ### Bugfix
 - PSR-0 via composer broken - @Progi1984 GH-51